Make local flag vars int
Sketch uses 270855 bytes (25%) of program storage space. Maximum is 1044464 bytes.
Global variables use 27940 bytes (34%) of dynamic memory, leaving 53980 bytes for local variables. Maximum is 81920 bytes.
./xtensa-lx106-elf/bin/xtensa-lx106-elf-objdump -t  -j .text1 /tmp/arduino_build_9615/*elf | sort -k1 | head -20
401000cc l     F .text1	00000014 twi_delay
401000ec l     F .text1	00000020 twi_reply$part$1
4010010c g     F .text1	00000035 twi_reply
4010014c g     F .text1	00000052 twi_stop
401001a0 g     F .text1	0000003b twi_releaseBus
40100204 g     F .text1	000001e6 twi_onTwipEvent
40100404 l     F .text1	000001f7 onSdaChange
40100608 l     F .text1	000002fd onSclChange
40100908 l     F .text1	0000003b onTimer

+++ cores/esp8266/core_esp8266_si2c.cpp
@@ -524,8 +524,8 @@ static void eventTask(ETSEvent *e)
 
 void ICACHE_RAM_ATTR onSclChange(void)
 {
-	uint8_t sda;
-	uint8_t scl;
+	unsigned int sda;
+	unsigned int scl;
 
 	sda	= SDA_READ();
 	scl = SCL_READ();
@@ -682,8 +682,8 @@ void ICACHE_RAM_ATTR onSclChange(void)
 
 void ICACHE_RAM_ATTR onSdaChange(void)
 {
-	uint8_t sda;
-	uint8_t scl;
+	unsigned int sda;
+	unsigned int scl;
 	sda	= SDA_READ();
 	scl = SCL_READ();
